requirements.txt
requests >= 2.28.2
google-api-python-client >= 2.109.0
google-auth-httplib2 >= 0.1.1
google-auth-oauthlib >= 1.1.0

